refactor(task02) divided the components to unique files and use arrow function
