+++ day-01/src/button.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Button = (props) => {
+  return (
+    <button onClick={props.onclick}>
+      {props.label}
+    </button>
+  )
+};
+
+Button.propTypes = {
+  label: PropTypes.string.isRequired,
+  onclick: PropTypes.func.isRequired
+};
+
+export default Button;
