+++ day-01/src/display.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Display = (props) => {
+  return (
+    <div>
+      {props.counter}
+    </div>
+  )
+};
+
+Display.propTypes = {
+  counter: PropTypes.number.isRequired
+};
+
+export default Display;
