+++ day-01/src/index.js
@@ -8,3 +8,4 @@ ReactDOM.render(
   document.getElementById('root'),
 );
 
+
